Fix tests and rename products parser file
