+++ web/app/integration-manager/merlins-connector/categories/parser.test.js
@@ -1,9 +1,9 @@
-/* eslint-env jquery, jest */
+/* eslint-env jquery, jest, node*/
 import {jquerifyHtmlFile} from 'progressive-web-sdk/dist/test-utils'
 import categoryProductsParser from './parser'
 
 describe('the product list parser', () => {
-    const $content = jquerifyHtmlFile('app/integration-manager/merlins-connector/categories/parser.test.html')
+    const $content = jquerifyHtmlFile(`${__dirname}/parser.test.html`)
     const parsedContent = categoryProductsParser($, $content)
 
     it('should extract the product list content from the rendered HTML', () => {
