+++ web/app/integration-manager/merlins-connector/commands.js
@@ -7,7 +7,7 @@ import {checkoutShippingParser, parseCheckoutData} from './checkout/parsers'
 import homeParser from './home/parser'
 import {parseNavigation} from './navigation/parser'
 import categoryProductsParser from './categories/parser'
-import {productListParser, productDetailsParser} from './products/parser'
+import {productListParser, productDetailsParser} from './products/parsers'
 import * as responses from './../responses'
 import {getCustomerEntityID} from '../../store/checkout/selectors'
 import {getIsLoggedIn} from '../../containers/app/selectors'
