+++ web/app/integration-manager/merlins-connector/products/parser.test.js
@@ -1,12 +1,12 @@
-/* eslint-env jquery, jest */
+/* eslint-env jquery, jest, node */
 import {jquerifyHtmlFile} from 'progressive-web-sdk/dist/test-utils'
 import {isURL} from 'validator'
-import {productDetailsParser, productListParser} from './parser'
+import {productDetailsParser, productListParser} from './parsers'
 
 /* eslint-disable max-nested-callbacks */
 
 describe('the ProductDetails product parser', () => {
-    const $content = jquerifyHtmlFile('app/containers/product-details/parsers/product-details-example.html')
+    const $content = jquerifyHtmlFile(`${__dirname}/product-details.test.html`)
     const parsedContent = productDetailsParser($, $content)
 
     it('extracts the title from the page', () => {
@@ -36,7 +36,7 @@ describe('the ProductDetails product parser', () => {
 })
 
 describe('the ProductList product parser', () => {
-    const $content = jquerifyHtmlFile('app/store/categories/parsers/product-list.test.html')
+    const $content = jquerifyHtmlFile(`${__dirname}/product-list.test.html`)
     const parsedContent = productListParser($, $content)
 
     it('should extract the product list content from the rendered HTML', () => {
